refactor(printutils): extract a shared spool backend from lpr/lpq/lpd
printutils.go combined three CLI fronts (lpr, lpq, lpd), queue
persistence (enqueue, nextID, writeControl, readQueue), and spool
draining (printJob) in one file and one Command type, so queue storage
policy and CLI policy were tightly coupled and partial-spool behavior was
hard to reason about beside the front-ends.

Extract a spool type (spool.go) that owns the queue storage policy -
control/data files, id allocation, list, and drain - behind init,
enqueue, list, and print methods. lpr (lpr.go), lpq (lpq.go), and lpd
(lpd.go) become thin front-ends that only parse flags and render output
over that backend. printutils.go keeps just the Command type and the
name-based dispatch tables. The local-only, no-network semantics are
unchanged.

Add a shell-level round-trip E2E (lpr a file plus stdin, lpq lists both,
lpd drains, queue empties) and unit tests for corrupt control files,
a missing data file, and a partial drain failure. Statement coverage
rises from 72.5% to 78.0%.

Closes #942
